requirements.txt
Django==5.2.1
Pillow==11.2.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
psycopg2-binary==2.9.9
